Viet truong

+++ back-end/app/Http/Controllers/Api/ResetPasswordController.php
@@ -16,38 +16,56 @@ class ResetPasswordController extends Controller
     //send mail
     public function sendMail(Request $request)
     {
-        $user = User::where('email', $request->email)->firstOrFail();
+        $user = User::where('email', $request->email)->first();
 
-        $passwordReset = PasswordReset::create([
-            'email' => $user->email,
-            'token' => Str::random(60),
-        ]);
+        if(!$user = User::where('email', $request->email)->first()) {
+            return response()->json([
+                'message' => 'Invalid Mail',
+                'type' => 'error'
+            ]);
+        }
+
+        $passwordReset = PasswordReset::updateOrCreate(
+            ['email' => $user->email],
+            ['token' => Str::random(60)]
+        );
 
         if ($passwordReset) {
             Mail::to($request->email)->send(new ResetPasswordMail($passwordReset->token));
         }
 
         return response()->json([
-            'message' => 'We have e-mailed your password reset token!'
+            'message' => 'We have e-mailed your password reset token!',
+            'type' => 'success'
         ]);
     }
 
-    public function resetPassword(Request $request, $token)
+    public function resetPassword(Request $request)
     {
-        $passwordReset = PasswordReset::where('token', $token)->firstOrFail();
+        $passwordReset = PasswordReset::where('token', $request->token)->first();
+
+        if($passwordReset == NULL) {
+            return response()->json([
+                'message' => 'This password reset token is invalid.',
+                'type' => 'error'
+            ]);
+        }
+
         if (Carbon::parse($passwordReset->updated_at)->addMinutes(720)->isPast()) {
             $passwordReset->delete();
 
             return response()->json([
                 'message' => 'This password reset token is invalid.',
-            ], 422);
+                'type' => 'error'
+            ]);
         }
         $user = User::where('email', $passwordReset->email)->firstOrFail();
-        $user->update($request->only('password'));
+        $user->update(['password' => bcrypt($request->input('password'))]);
         $passwordReset->delete();
 
         return response()->json([
             'message' => 'Change password was successfully',
+            'type' => 'success',
             'user' => $user
         ]);
     }
